Fix: Arrumando várias coisas

+++ Gerenciamento_Estoque_Desafio_3/src/main/java/estoque/desafio/gerenciamento/repositories/ProjetoRepository.java
@@ -3,6 +3,6 @@ package estoque.desafio.gerenciamento.repositories;
 import estoque.desafio.gerenciamento.entities.Projeto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjetoRepository extends JpaRepository<Projeto, Integer> {
+public interface ProjetoRepository extends JpaRepository<Projeto, Long> {
 
 }
